requirements.txt
flask
sqlalchemy
flask-sqlalchemy
marshmallow
flask-smorest
python-dotenv
flask-jwt-extended
passlib
flask-migrate
cloudinary
pillow
numpy
opencv-python
tensorflow
psycopg2
flask-cors
authlib
onnxruntime
rembg
scikit-learn





